+++ R/MANOVA.RM-package.R
@@ -1,7 +1,7 @@
 #'MANOVA.RM: A package for calculating test statistics and their resampling versions for
 #'heteroscedastic semi-parametric multivariate data or repeated measures designs.
 #'
-#'The MANOVA.RM package provides two important functions: MANOVA() and RM() which 
+#'The MANOVA.RM package provides three important functions: MANOVA(), RM() and multRM() which 
 #'will be explained in detail below.
 #'
 #'@section MANOVA and MANOVA.wide function: The MANOVA() and MANOVA.wide() functions provide
@@ -29,7 +29,15 @@
 #'  resampling approaches (Permutation, parametric bootstrap, Wild bootstrap).
 #'  For further details, see \code{RM}.
 #'  
-#'@references Konietschke, F., Bathke, A. C., Harrar, S. W. and Pauly, M. (2015).
+#'@section multRM function: The multRM() function is a combination of the procedures
+#'  above suited for multivariate repeated measures designs. It provides the WTS and the MATS
+#'  along with p-values based on a parametric or a wild bootstrap approach.
+#'  
+#'@references Friedrich, S., Konietschke, F., and Pauly, M. (2019). Resampling-Based Analysis
+#' of Multivariate Data and Repeated Measures Designs with the R Package MANOVA.RM. 
+#' The R Journal, 11(2), 380-400.
+#'
+#'Konietschke, F., Bathke, A. C., Harrar, S. W. and Pauly, M. (2015).
 #'  Parametric and nonparametric bootstrap methods for general MANOVA. Journal
 #'  of Multivariate Analysis, 140, 291-301.
 #'