+++ inst/NEWS.md
@@ -1,3 +1,9 @@
+# MANOVA.RM 0.4.1
+* include new function multRM for multivariate repeated measures designs
+* update documentation and vignette
+* update output of RM (now also states the names of the within-subject factors)
+* improve post-hoc comparisons
+
 # MANOVA.RM 0.3.4
 * include warning for incorrectly labelled subjects (in RM designs with more than one whole-plot factor)
 * include graphics for example documentation